# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResumeGPT AI Enterprise API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface ImportWorkspaceProfileInput {
  profileUrl: string;
  /** @minLength 8 */
  model?: string;
}

export interface ProfileImportInput {
  /** @minLength 8 */
  profileUrl: string;
}

export type ProfileSaveInputProfile = { [key: string]: unknown };

export interface ProfileSaveInput {
  /** @nullable */
  profileUrl?: string | null;
  profile: ProfileSaveInputProfile;
}

export interface JobAnalyzeInput {
  /** @nullable */
  jobUrl: string | null;
  /**
     * @maxLength 50000
     * @nullable
     */
  jobDescription: string | null;
}

export type ResumeGenerateInputMode = typeof ResumeGenerateInputMode[keyof typeof ResumeGenerateInputMode];


export const ResumeGenerateInputMode = {
  manual: 'manual',
  auto: 'auto',
  guided: 'guided',
  expert: 'expert',
} as const;

/**
 * @nullable
 */
export type ResumeGenerateInputManualEdits = { [key: string]: unknown } | null;

export interface ResumeGenerateInput {
  jobAnalysisId: number;
  mode: ResumeGenerateInputMode;
  model?: string;
  /** @minLength 1 */
  templateId: string;
  /** @nullable */
  manualEdits?: ResumeGenerateInputManualEdits;
}

export type ProfileRecordProfile = { [key: string]: unknown };

export interface ProfileRecord {
  id: number;
  workspaceKey: string;
  profileUrl: string;
  source: string;
  profile: ProfileRecordProfile;
  fetchedAt: string;
}

export type JobAnalysisRecordSource = typeof JobAnalysisRecordSource[keyof typeof JobAnalysisRecordSource];


export const JobAnalysisRecordSource = {
  url: 'url',
  description: 'description',
} as const;

export type JobAnalysisRecordJob = { [key: string]: unknown };

export type JobAnalysisRecordComparison = { [key: string]: unknown };

export interface JobAnalysisRecord {
  id: number;
  workspaceKey: string;
  /** @nullable */
  jobUrl?: string | null;
  source: JobAnalysisRecordSource;
  job: JobAnalysisRecordJob;
  comparison: JobAnalysisRecordComparison;
  analyzedAt: string;
}

export type ResumeVersionRecordMode = typeof ResumeVersionRecordMode[keyof typeof ResumeVersionRecordMode];


export const ResumeVersionRecordMode = {
  manual: 'manual',
  auto: 'auto',
  guided: 'guided',
  expert: 'expert',
} as const;

export type ResumeVersionRecordResume = { [key: string]: unknown };

export interface ResumeVersionRecord {
  id: number;
  workspaceKey: string;
  jobAnalysisId: number;
  mode: ResumeVersionRecordMode;
  templateId: string;
  resume: ResumeVersionRecordResume;
  createdAt: string;
}

export interface WorkspaceState {
  profile: ProfileRecord | null;
  job: JobAnalysisRecord | null;
  resume: ResumeVersionRecord | null;
}

/**
 * Error response
 */
export type ErrorResponse = Error;

export type WorkspaceKeyParameter = string;